VFB-415: refactored code
