+++ src/app/clients/getClientParcelsData.ts
@@ -3,7 +3,6 @@ import { DatabaseError } from "@/app/errorClasses";
 import { Data } from "@/components/DataViewer/DataViewer";
 import { logErrorReturnLogId } from "@/logger/logger";
 import { formatDatetimeAsDate } from "@/common/format";
-import { ex } from "@fullcalendar/core/internal-common";
 
 export type RawClientParcelsDetails = Awaited<ReturnType<typeof getRawClientParcelsDetails>>;
 
@@ -55,12 +54,6 @@ export interface ExpandedClientParcelDetails extends Data {
     collectionCentre: string;
 }
 
-export interface ExpandedClientParcelStats extends Data {
-    totalParcels: number;
-    totalSuccessful: number;
-    lastSixMonthsSuccessful: number;
-}
-
 export const rawDataToClientParcelsDetails = (
     parcel: ClientParcelDetails
 ): ExpandedClientParcelDetails => {
@@ -72,51 +65,3 @@ export const rawDataToClientParcelsDetails = (
     };
 };
 
-
-export const getClientParcelsStats = async (clientId: string): Promise<ExpandedClientParcelStats[]> => {
-    const parcelsData: ExpandedClientParcelDetails[] = await getClientParcelsDetails(clientId);
-    let parcelIdList: string[] = parcelsData.map((clientDetails) => clientDetails.parcelId);
-    // console.log(parcelsData);
-
-    const { data, error } = await supabase
-        .from("parcels_events")
-        .select(
-            `
-            last_event_timestamp,
-            all_events
-        `
-        )
-        .in("parcel_id", parcelIdList);
-
-    if (error) {
-        const logId = await logErrorReturnLogId("Error with fetch: Client parcels details", error);
-        throw new DatabaseError("fetch", "client parcels", logId);
-    }
-    // console.log(data.length) // total parcels
-    let deliveredParcels = data.filter((parcel) =>
-        parcel.all_events?.includes("Delivered") ||
-        parcel.all_events?.includes("Parcel Collected") ||
-        parcel.all_events?.includes("Fulfilled with Trussell")
-    );
-    // console.log(deliveredParcels.length); // total successful parcels
-
-    const todayTime = new Date().getTime();
-    let sixMonthsAgo = new Date();
-    sixMonthsAgo.setMonth(new Date().getMonth() - 6);
-    const sixMonthsAgoTime = sixMonthsAgo.getTime();
-
-    let deliveredLastSixMonths = deliveredParcels.filter((parcel) => {
-        if(!parcel.last_event_timestamp) return false;
-        let time = Date.parse(parcel.last_event_timestamp);
-        return time >= sixMonthsAgoTime && time <= todayTime;
-        }
-    );
-    // console.log(deliveredLastSixMonths.length);
-    
-    return [{
-        totalParcels: data.length,
-        totalSuccessful: deliveredParcels.length,
-        lastSixMonthsSuccessful: deliveredLastSixMonths.length
-    }];
-
-}
